refactor: split star disk material into modules
